+++ test/core/buffer_test.hpp
@@ -9,8 +9,8 @@
 #define BEAST_TEST_BUFFER_TEST_HPP
 
 #include <beast/core/string_view.hpp>
+#include <beast/core/read_size.hpp>
 #include <beast/core/type_traits.hpp>
-#include <beast/core/detail/read_size_helper.hpp>
 #include <beast/core/detail/type_traits.hpp>
 #include <boost/asio/buffer.hpp>
 #include <algorithm>
@@ -110,37 +110,6 @@ size_rev_post(ConstBufferSequence const& buffers)
         n += boost::asio::buffer_size(*it);
     }
     return n;
-}
-
-namespace has_read_size_helper
-{
-    using beast::detail::read_size_helper;
-
-    template<class T, class = void>
-    struct trait : std::false_type {};
-
-    template<class T>
-    struct trait<T, decltype(
-        std::declval<std::size_t&>() =
-            read_size_helper(std::declval<T>(), 0),
-        (void)0)> : std::true_type
-    {
-    };
-}
-
-//------------------------------------------------------------------------------
-
-// Make sure read_size_helper works
-template<class DynamicBuffer>
-inline
-void
-check_read_size_helper()
-{
-    static_assert(is_dynamic_buffer<DynamicBuffer>::value,
-        "DynamicBuffer requirements not met ");
-
-    static_assert(has_read_size_helper::trait<DynamicBuffer>::value,
-        "Missing read_size_helper for dynamic buffer");
 }
 
 } // test