+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 55:
 * Don't allocate memory to handle obs-fold
 * Add Beast CMake interface target
 * Avoid a parser allocation using non-flat buffer
+* read_size replaces read_size_helper
 
 --------------------------------------------------------------------------------